requirements.txt
fastapi
pymongo
python-dotenv
pydantic
uvicorn
pydantic==1.8
pytest
httpx
requests